add docs 7

+++ composeApp/src/commonMain/kotlin/ramble/sokol/tgm_inverse/FinishGameScreen.kt
@@ -49,7 +49,7 @@ class FinishGameScreen(
     val userEntityCreate: UserEntityCreate,
     val bodyUserCreate: UserEntityCreateResponse,
     val nameSong: String,
-    val reward: Int,
+    val reward: Long,
     val countStar: Int
 ) : Screen {
